clients/badge: configure organization default message in onboarding and settings
