+++ clients/apps/web/src/components/Dashboard/BadgeMessageForm.tsx
@@ -1,19 +1,18 @@
 import { Badge } from '@/../../../packages/polarkit/src/components'
 import { Marked } from '@ts-stack/markdown'
 import { useTheme } from 'next-themes'
-import { IssueDashboardRead } from 'polarkit/api/client'
 import { classNames } from 'polarkit/utils'
 import { ChangeEvent, MouseEvent, useEffect, useRef, useState } from 'react'
 import LabeledRadioButton from '../UI/LabeledRadioButton'
 
 const BadgeMessageForm = (props: {
   orgName: string
-  repoName: string
-  issue: IssueDashboardRead
-  onBadgeWithComment: (comment: string) => Promise<void>
+  value: string
+  onUpdate: (comment: string) => Promise<void>
+  showUpdateButton: boolean
+  onChange: (comment: string) => void
+  innerClassNames: string
 }) => {
-  // TODO: Get this from backend based on settings to reflect how it's currently embedded.
-  // Starting with empty string now since that is how it currently would work.
   const [message, setMessage] = useState('')
 
   const [descriptionMode, setDescirptionMode] = useState('View')
@@ -22,22 +21,23 @@ const BadgeMessageForm = (props: {
 
   useEffect(() => {
     if (ref.current) {
-      ref.current.innerHTML = Marked.parse(message)
+      ref.current.innerHTML = Marked.parse(props.value)
     }
-  }, [ref, message, descriptionMode])
+    setMessage(props.value)
+  }, [ref, props.value, descriptionMode])
 
   const [canSave, setCanSave] = useState(false)
 
   const onChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
     setMessage(e.target.value)
-    setCanSave(e.target.value !== props.issue.badge_custom_content)
+    setCanSave(e.target.value !== props.value)
   }
 
   const [isLoading, setIsLoading] = useState(false)
 
   const onClickUpdate = async (e: MouseEvent<HTMLButtonElement>) => {
     setIsLoading(true)
-    await props.onBadgeWithComment(message)
+    await props.onUpdate(message)
     setIsLoading(false)
   }
 
@@ -47,7 +47,7 @@ const BadgeMessageForm = (props: {
     <div className="flex flex-col space-y-3">
       <div className="text-gray flex items-center justify-between">
         <div className="text-sm font-medium">
-          Customize Polar description in issue
+          Markdown added to the end of the issue description
         </div>
         <LabeledRadioButton
           values={['View', 'Edit']}
@@ -55,7 +55,12 @@ const BadgeMessageForm = (props: {
           onSelected={setDescirptionMode}
         />
       </div>
-      <div className="rounded-xl bg-white py-3.5 px-5 shadow dark:bg-gray-800 dark:ring-1 dark:ring-gray-600">
+      <div
+        className={classNames(
+          props.innerClassNames,
+          'rounded-xl bg-white py-3.5 px-5 dark:bg-gray-800 dark:ring-1 dark:ring-gray-600',
+        )}
+      >
         {descriptionMode === 'View' && (
           <>
             <div className="prose dark:prose-invert" ref={ref} />
@@ -81,19 +86,21 @@ const BadgeMessageForm = (props: {
           Template variables: <code>{'{badge}'}</code>, <code>{'{repo}'}</code>
         </div> */}
         <div className="flex-1"></div>
-        <button
-          onClick={onClickUpdate}
-          disabled={!canSave}
-          className={classNames(
-            isLoading ? 'cursor-wait' : '',
-            canSave
-              ? 'cursor-pointer text-blue-600 dark:text-blue-500'
-              : 'text-gray-400 dark:text-gray-500',
-            'text-sm font-medium',
-          )}
-        >
-          Update
-        </button>
+        {props.showUpdateButton && (
+          <button
+            onClick={onClickUpdate}
+            disabled={!canSave}
+            className={classNames(
+              isLoading ? 'cursor-wait' : '',
+              canSave
+                ? 'cursor-pointer text-blue-600 dark:text-blue-500'
+                : 'text-gray-400 dark:text-gray-500',
+              'text-sm font-medium',
+            )}
+          >
+            Update
+          </button>
+        )}
       </div>
     </div>
   )
