+++ clients/packages/polarkit/src/api/client/models/OrganizationBadgeSettingsRead.ts
@@ -6,6 +6,7 @@ import type { RepositoryBadgeSettingsRead } from './RepositoryBadgeSettingsRead'
 
 export type OrganizationBadgeSettingsRead = {
   show_amount: boolean;
+  message?: string;
   repositories: Array<RepositoryBadgeSettingsRead>;
 };
