+++ clients/packages/polarkit/src/api/client/models/OrganizationPrivateRead.ts
@@ -26,6 +26,7 @@ export type OrganizationPrivateRead = {
   installation_updated_at?: string;
   installation_suspended_at?: string;
   onboarded_at?: string;
+  default_badge_custom_content?: string;
   id: string;
   status: Status;
   created_at: string;
